move dropdown title component
